predadorpresa/requirements.txt
backcall==0.2.0
cycler==0.10.0
decorator==5.0.6
importlib-metadata==3.10.0
importlib-resources==5.1.2
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
modsimpy==1.1.3
mpmath==1.2.1
numpy==1.19.5
packaging==20.9
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
Pint==0.17
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
sympy==1.7.1
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
zipp==3.4.1
